refactor: Improve performance at nitrogen balance calculation for ammonia emissions of fertilizer application by interating only once over each fertilizer application
