+++ .changeset/ten-garlics-marry.md
@@ -1,0 +1,5 @@
+---
+"@svenvw/fdm-calculator": patch
+---
+
+Improve performance at nitrogen balance calculation for ammonia emissions of fertilizer application by interating only once over each fertilizer application
